organize types: ServiceLevel
