+++ types/ServiceLevel/APC.d.ts
@@ -1,0 +1,12 @@
+export declare type ServiceLevelAPC =
+| 'parcelConnectBookService'
+| 'parcelConnectExpeditedDDP'
+| 'parcelConnectExpeditedDDU'
+| 'parcelConnectPriorityDDP'
+| 'parcelConnectPriorityDDPDelcon'
+| 'parcelConnectPriorityDDU'
+| 'parcelConnectPriorityDDUDelcon'
+| 'parcelConnectPriorityDDUPQW'
+| 'parcelConnectStandardDDU'
+| 'parcelConnectStandardDDUPQW'
+| 'parcelConnectPacketDDU';
