+++ types/ServiceLevel/AmazonMws.d.ts
@@ -1,0 +1,10 @@
+export declare type ServiceLevelAmazonMws =
+  | 'UPS Rates'
+  | 'USPS Rates'
+  | 'FedEx Rates'
+  | 'UPS Labels'
+  | 'USPS Labels'
+  | 'FedEx Labels'
+  | 'UPS Tracking'
+  | 'USPS Tracking'
+  | 'FedEx Tracking';
