+++ types/ServiceLevel/Aramex.d.ts
@@ -1,0 +1,10 @@
+export declare type ServiceLevelAramex =
+| 'Domestic'
+| 'PriorityDocumentExpress'
+| 'PriorityParcelExpress'
+| 'PriorityLetterExpress'
+| 'DeferredDocumentExpress'
+| 'DeferredParcelExpress'
+| 'GroundDocumentExpress'
+| 'GroundParcelExpress'
+| 'EconomyParcelExpress';
