+++ types/ServiceLevel/ArrowXL.d.ts
@@ -1,0 +1,28 @@
+export declare type ServiceLevelArrowXL =
+| '48HRBrownInstall'
+| '48HRBrownInstallRemove'
+| '48HROther'
+| '48HROtherRemove'
+| '48HRWhiteInstall'
+| '48HRWhiteInstallRemove'
+| '48HRWhiteRemove'
+| 'ExpressBrownInstall'
+| 'ExpressBrownInstallRemove'
+| 'ExpressOther'
+| 'ExpressOtherRemove'
+| 'ExpressWhiteInstall'
+| 'ExpressWhiteInstallRemove'
+| 'ExpressWhiteRemove'
+| 'StandardBrownInstall'
+| 'StandardBrownInstallRemove'
+| 'StandardBrownRemove'
+| 'StandardOther'
+| 'StandardOtherAssemble'
+| 'StandardOtherNoUnpack'
+| 'StandardOtherRemove'
+| 'StandardOtherUnpack'
+| 'StandardOtherUnpackLeavePackaging'
+| 'StandardReturns'
+| 'StandardWhiteInstall'
+| 'StandardWhiteInstallRemove'
+| 'StandardWhiteRemove';
