+++ types/ServiceLevel/Asendia.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelAsendia = 'PMI' | 'ePacket' | 'IPA' | 'ISAL';
