+++ types/ServiceLevel/AustraliaPost.d.ts
@@ -1,0 +1,7 @@
+export declare type ServiceLevelAustraliaPost =
+| 'ExpressPost'
+| 'ExpressPostSignature'
+| 'ParcelPost'
+| 'ParcelPostSignature'
+| 'ParcelPostExtra'
+| 'ParcelPostWinePlusSignature';
