+++ types/ServiceLevel/AxlehireV3.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelAxlehireV3 = 'AxleHireDelivery';
