+++ types/ServiceLevel/BorderGuru.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelBorderGuru = 'ECONOMY' | 'PRIORITY' | 'EXPEDITED';
