+++ types/ServiceLevel/CanadaPost.d.ts
@@ -1,0 +1,24 @@
+export declare type ServiceLevelCanadaPost =
+| 'RegularParcel'
+| 'ExpeditedParcel'
+| 'Xpresspost'
+| 'XpresspostCertified'
+| 'Priority'
+| 'LibraryBooks'
+| 'ExpeditedParcelUSA'
+| 'PriorityWorldwideEnvelopeUSA'
+| 'PriorityWorldwidePakUSA'
+| 'PriorityWorldwideParcelUSA'
+| 'SmallPacketUSAAir'
+| 'TrackedPacketUSA'
+| 'TrackedPacketUSALVM'
+| 'XpresspostUSA'
+| 'XpresspostInternational'
+| 'InternationalParcelAir'
+| 'InternationalParcelSurface'
+| 'PriorityWorldwideEnvelopeIntl'
+| 'PriorityWorldwidePakIntl'
+| 'PriorityWorldwideParcelIntl'
+| 'SmallPacketInternationalAir'
+| 'SmallPacketInternationalSurface'
+| 'TrackedPacketInternational';
