+++ types/ServiceLevel/Canpar.d.ts
@@ -1,0 +1,13 @@
+export declare type ServiceLevelCanpar =
+| 'Ground'
+| 'SelectLetter'
+| 'SelectPak'
+| 'Select'
+| 'OvernightLetter'
+| 'OvernightPak'
+| 'Overnight'
+| 'SelectUSA'
+| 'USAPak'
+| 'USALetter'
+| 'USA'
+| 'International';
