+++ types/ServiceLevel/ColisPrive.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelColisPrive = 'Parcel';
