+++ types/ServiceLevel/ColumbusLastMile.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelColumbusLastMile = 'DISTRIBUTION' | 'Same Day';
