+++ types/ServiceLevel/CouriersPlease.d.ts
@@ -1,0 +1,32 @@
+export declare type ServiceLevelCouriersPlease =
+| 'DomesticPrioritySignature'
+| 'DomesticPriority'
+| 'DomesticOffPeakSignature'
+| 'DomesticOffPeak'
+| 'GoldDomesticSignature'
+| 'GoldDomestic'
+| 'AustralianCityExpressSignature'
+| 'AustralianCityExpress'
+| 'DomesticSaverSignature'
+| 'DomesticSaver'
+| 'RoadExpress'
+| '5KgSatchel'
+| '3KgSatchel'
+| '1KgSatchel'
+| '5KgSatchelATL'
+| '3KgSatchelATL'
+| '1KgSatchelATL'
+| '500GramSatchel'
+| '500GramSatchelATL'
+| '25KgParcel'
+| '10KgParcel'
+| '5KgParcel'
+| '3KgParcel'
+| '1KgParcel'
+| '500GramParcel'
+| '500GramParcelATL'
+| 'ExpressInternationalPriority'
+| 'InternationalSaver'
+| 'InternationalExpressImport'
+| 'ExpressInternationalPriority'
+| 'InternationalExpress';
