+++ types/ServiceLevel/DHLEcommerceAsia.d.ts
@@ -1,0 +1,5 @@
+export declare type ServiceLevelDHLEcommerceAsia =
+| 'Packet'
+| 'PacketPlus'
+| 'ParcelDirect'
+| 'ParcelDirectExpedited';
