+++ types/ServiceLevel/DHLExpress.d.ts
@@ -1,0 +1,33 @@
+export declare type ServiceLevelDHLExpress =
+| 'BreakBulkEconomy'
+| 'BreakBulkExpress'
+| 'DomesticEconomySelect'
+| 'DomesticExpress'
+| 'DomesticExpress1030'
+| 'DomesticExpress1200'
+| 'EconomySelect'
+| 'EconomySelectNonDoc'
+| 'EuroPack'
+| 'EuropackNonDoc'
+| 'Express1030'
+| 'Express1030NonDoc'
+| 'Express1200NonDoc'
+| 'Express1200'
+| 'Express900'
+| 'Express900NonDoc'
+| 'ExpressEasy'
+| 'ExpressEasyNonDoc'
+| 'ExpressEnvelope'
+| 'ExpressWorldwide'
+| 'ExpressWorldwideB2C'
+| 'ExpressWorldwideB2CNonDoc'
+| 'ExpressWorldwideECX'
+| 'ExpressWorldwideNonDoc'
+| 'FreightWorldwide'
+| 'GlobalmailBusiness'
+| 'JetLine'
+| 'JumboBox'
+| 'LogisticsServices'
+| 'SameDay'
+| 'SecureLine'
+| 'SprintLine';
