+++ types/ServiceLevel/DPD.d.ts
@@ -1,0 +1,10 @@
+export declare type ServiceLevelDPD =
+| 'DPDCLASSIC'
+| 'DPD8:30'
+| 'DPD10:00'
+| 'DPD12:00'
+| 'DPD18:00'
+| 'DPDEXPRESS'
+| 'DPDPARCELLETTER'
+| 'DPDPARCELLETTERPLUS'
+| 'DPDINTERNATIONALMAIL';
