+++ types/ServiceLevel/DPDUK.d.ts
@@ -1,0 +1,47 @@
+export declare type ServiceLevelDPDUK =
+| 'AirExpressInternationalAir'
+| 'AirClassicInternationalAir'
+| 'ParcelSunday'
+| 'FreightParcelSunday'
+| 'PalletSunday'
+| 'PalletDpdClassic'
+| 'ExpresspakDpdClassic'
+| 'ExpresspakSunday'
+| 'ParcelDpdClassic'
+| 'ParcelDpdTwoDay'
+| 'ParcelDpdNextDay'
+| 'ParcelDpd12'
+| 'ParcelDpd10'
+| 'ParcelDpd10'
+| 'ParcelReturnToShop'
+| 'ParcelSaturday'
+| 'ParcelSaturday12'
+| 'ParcelSaturday10'
+| 'ParcelSaturday10'
+| 'ParcelSunday12'
+| 'FreightParcelDpdClassic'
+| 'FreightParcelSunday12'
+| 'ExpresspakDpdNextDay'
+| 'ExpresspakDpd12'
+| 'ExpresspakDpd10'
+| 'ExpresspakDpd10'
+| 'ExpresspakSaturday'
+| 'ExpresspakSaturday12'
+| 'ExpresspakSaturday10'
+| 'ExpresspakSaturday10'
+| 'ExpresspakSunday12'
+| 'PalletSunday12'
+| 'PalletDpdTwoDay'
+| 'PalletDpdNextDay'
+| 'PalletDpd12'
+| 'PalletDpd10'
+| 'PalletSaturday'
+| 'PalletSaturday12'
+| 'PalletSaturday10'
+| 'FreightParcelDpdTwoDay'
+| 'FreightParcelDpdNextDay'
+| 'FreightParcelDpd12'
+| 'FreightParcelDpd10'
+| 'FreightParcelSaturday'
+| 'FreightParcelSaturday12'
+| 'FreightParcelSaturday10';
