+++ types/ServiceLevel/DaiPost.d.ts
@@ -1,0 +1,5 @@
+export declare type ServiceLevelDaiPost =
+| 'DomesticTracked'
+| 'InternationalEconomy'
+| 'InternationalStandard'
+| 'InternationalExpress';
