+++ types/ServiceLevel/Deliv.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelDeliv = 'Scheduled' | 'OnDemand';
