+++ types/ServiceLevel/DeutschePost.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelDeutschePost = 'PacketPlus';
