+++ types/ServiceLevel/DeutschePostUK.d.ts
@@ -1,0 +1,7 @@
+export declare type ServiceLevelDeutschePostUK =
+| 'PriorityPacketPlus'
+| 'PriorityPacket'
+| 'PriorityPacketTracked'
+| 'BusinessMailRegistered'
+| 'StandardPacket'
+| 'BusinessMailStandard';
