+++ types/ServiceLevel/DhlEcs.d.ts
@@ -1,0 +1,11 @@
+export declare type ServiceLevelDhlEcs =
+| 'DHLParcelExpedited'
+| 'DHLParcelExpeditedMax'
+| 'DHLParcelGround'
+| 'DHLBPMExpedited'
+| 'DHLBPMGround'
+| 'DHLParcelInternationalDirect'
+| 'DHLParcelInternationalStandard'
+| 'DHLPacketInternational'
+| 'DHLParcelInternationalDirectPriority'
+| 'DHLParcelInternationalDirectStandard';
