+++ types/ServiceLevel/Dicom.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelDicom = 'Ground';
