+++ types/ServiceLevel/Estafeta.d.ts
@@ -1,0 +1,7 @@
+export declare type ServiceLevelEstafeta =
+| 'NextDayBy930'
+| 'NextDayBy1130'
+| 'NextDay'
+| 'Ground'
+| 'TwoDay'
+| 'LTL';
