+++ types/ServiceLevel/Fastway.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelFastway = 'Parcel' | 'Satchel';
