+++ types/ServiceLevel/FedEx.d.ts
@@ -1,0 +1,13 @@
+export declare type ServiceLevelFedEx =
+| 'FEDEX_GROUND'
+| 'FEDEX_2_DAY'
+| 'FEDEX_2_DAY_AM'
+| 'FEDEX_EXPRESS_SAVER'
+| 'STANDARD_OVERNIGHT'
+| 'FIRST_OVERNIGHT'
+| 'PRIORITY_OVERNIGHT'
+| 'INTERNATIONAL_ECONOMY'
+| 'INTERNATIONAL_FIRST'
+| 'INTERNATIONAL_PRIORITY'
+| 'GROUND_HOME_DELIVERY'
+| 'SMART_POST';
