+++ types/ServiceLevel/FedExCrossBorder.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelFedExCrossBorder = 'CBEC' | 'CBECL';
