+++ types/ServiceLevel/FedExSameDayCity.d.ts
@@ -1,0 +1,5 @@
+export declare type ServiceLevelFedExSameDayCity =
+| 'EconomyService'
+| 'StandardService'
+| 'PriorityService'
+| 'LastMile';
