+++ types/ServiceLevel/FedexSmartPost.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelFedexSmartPost = 'SMART_POST';
