+++ types/ServiceLevel/GSO.d.ts
@@ -1,0 +1,8 @@
+export declare type ServiceLevelGSO =
+| 'EarlyPriorityOvernight'
+| 'PriorityOvernight'
+| 'CaliforniaParcelService'
+| 'SaturdayDeliveryService'
+| 'EarlySaturdayService'
+| 'Ground'
+| 'Overnight';
