+++ types/ServiceLevel/Globegistics.d.ts
@@ -1,0 +1,17 @@
+export declare type ServiceLevelGlobegistics =
+| 'PMEI'
+| 'PMI'
+| 'eComDomestic'
+| 'eComEurope'
+| 'eComExpress'
+| 'eComExtra'
+| 'eComIPA'
+| 'eComISAL'
+| 'eComPMEIDutyPaid'
+| 'eComPMIDutyPaid'
+| 'eComPacket'
+| 'eComPacketDDP'
+| 'eComPriority'
+| 'eComStandard'
+| 'eComTrackedDDP'
+| 'eComTrackedDDU';
