+++ types/ServiceLevel/Hermes.d.ts
@@ -1,0 +1,5 @@
+export declare type ServiceLevelHermes =
+| 'DomesticDelivery'
+| 'DomesticDeliverySigned'
+| 'InternationalDelivery'
+| 'InternationalDeliverySigned';
