+++ types/ServiceLevel/InterlinkExpress.d.ts
@@ -1,0 +1,45 @@
+export declare type ServiceLevelInterlinkExpress =
+| 'InterlinkAirClassicInternationalAir'
+| 'InterlinkAirExpressInternationalAir'
+| 'InterlinkExpresspak1By10:30'
+| 'InterlinkExpresspak1By12'
+| 'InterlinkExpresspak1NextDay'
+| 'InterlinkExpresspak1Saturday'
+| 'InterlinkExpresspak1SaturdayBy10:30'
+| 'InterlinkExpresspak1SaturdayBy12'
+| 'InterlinkExpresspak1Sunday'
+| 'InterlinkExpresspak1SundayBy12'
+| 'InterlinkExpresspak5By10'
+| 'InterlinkExpresspak5By10:30'
+| 'InterlinkExpresspak5By12'
+| 'InterlinkExpresspak5NextDay'
+| 'InterlinkExpresspak5Saturday'
+| 'InterlinkExpresspak5SaturdayBy10'
+| 'InterlinkExpresspak5SaturdayBy10:30'
+| 'InterlinkExpresspak5SaturdayBy12'
+| 'InterlinkExpresspak5Sunday'
+| 'InterlinkExpresspak5SundayBy12'
+| 'InterlinkFreightBy10'
+| 'InterlinkFreightBy12'
+| 'InterlinkFreightNextDay'
+| 'InterlinkFreightSaturday'
+| 'InterlinkFreightSaturdayBy10'
+| 'InterlinkFreightSaturdayBy12'
+| 'InterlinkFreightSunday'
+| 'InterlinkFreightSundayBy12'
+| 'InterlinkParcelBy10'
+| 'InterlinkParcelBy10:30'
+| 'InterlinkParcelBy12'
+| 'InterlinkParcelDpdEuropeByRoad'
+| 'InterlinkParcelNextDay'
+| 'InterlinkParcelReturn'
+| 'InterlinkParcelReturnToShop'
+| 'InterlinkParcelSaturday'
+| 'InterlinkParcelSaturdayBy10'
+| 'InterlinkParcelSaturdayBy10:30'
+| 'InterlinkParcelSaturdayBy12'
+| 'InterlinkParcelShipToShop'
+| 'InterlinkParcelSunday'
+| 'InterlinkParcelSundayBy12'
+| 'InterlinkParcelTwoDay'
+| 'InterlinkPickupParcelDpdEuropeByRoad';
