+++ types/ServiceLevel/LSO.d.ts
@@ -1,0 +1,8 @@
+export declare type ServiceLevelLSO =
+| 'GroundEarly'
+| 'GroundBasic'
+| 'PriorityBasic'
+| 'PriorityEarly'
+| 'PrioritySaturday'
+| 'Priority2ndDay'
+| 'SameDay';
