+++ types/ServiceLevel/LaserShipV2.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelLaserShipV2 = 'SameDay' | 'NextDay' | 'Weekend';
