+++ types/ServiceLevel/Liefery.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelLiefery = 'TourAppointment20161124From1600To1900';
