+++ types/ServiceLevel/LoomisExpress.d.ts
@@ -1,0 +1,5 @@
+export declare type ServiceLevelLoomisExpress =
+| 'LoomisGround'
+| 'LoomisExpress1800'
+| 'LoomisExpress1200'
+| 'LoomisExpress900';
