+++ types/ServiceLevel/Newgistics.d.ts
@@ -1,0 +1,8 @@
+export declare type ServiceLevelNewgistics =
+| 'ParcelSelect'
+| 'ParcelSelectLightweight'
+| 'Ground'
+| 'Express'
+| 'FirstClassMail'
+| 'PriorityMail'
+| 'BoundPrintedMatter';
