+++ types/ServiceLevel/OmniParcel.d.ts
@@ -1,0 +1,5 @@
+export declare type ServiceLevelOmniParcel =
+| 'eCommerce Standard Tracked'
+| 'eCommerce Express Tracked'
+| 'Domestic Express'
+| 'Domestic Standard';
