+++ types/ServiceLevel/OnTrac.d.ts
@@ -1,0 +1,6 @@
+export declare type ServiceLevelOnTrac =
+| 'Sunrise'
+| 'Gold'
+| 'OnTracGround'
+| 'SameDay'
+| 'PalletizedFreight';
