+++ types/ServiceLevel/OnTracDirectPost.d.ts
@@ -1,0 +1,7 @@
+export declare type ServiceLevelOnTracDirectPost =
+| 'FirstClassMail'
+| 'PriorityMail'
+| 'BoundPrintedMatter'
+| 'MediaMail'
+| 'ParcelSelect'
+| 'ParcelSelectLightweight';
