+++ types/ServiceLevel/OsmWorldwide.d.ts
@@ -1,0 +1,10 @@
+export declare type ServiceLevelOsmWorldwide =
+| 'First'
+| 'Expedited'
+| 'ParcelSelectLightweight'
+| 'Priority'
+| 'BPM'
+| 'ParcelSelect'
+| 'MediaMail'
+| 'MarketingParcel'
+| 'MarketingParcelTracked';
