+++ types/ServiceLevel/Purolator.d.ts
@@ -1,0 +1,60 @@
+export declare type ServiceLevelPurolator =
+| 'PurolatorExpress'
+| 'PurolatorExpress12PM'
+| 'PurolatorExpressPack12PM'
+| 'PurolatorExpressBox12PM'
+| 'PurolatorExpressEnvelope12PM'
+| 'PurolatorExpress1030AM'
+| 'PurolatorExpress9AM'
+| 'PurolatorExpressBox'
+| 'PurolatorExpressBox1030AM'
+| 'PurolatorExpressBox9AM'
+| 'PurolatorExpressBoxEvening'
+| 'PurolatorExpressBoxInternational'
+| 'PurolatorExpressBoxInternational1030AM'
+| 'PurolatorExpressBoxInternational1200'
+| 'PurolatorExpressBoxInternational9AM'
+| 'PurolatorExpressBoxUS'
+| 'PurolatorExpressBoxUS1030AM'
+| 'PurolatorExpressBoxUS1200'
+| 'PurolatorExpressBoxUS9AM'
+| 'PurolatorExpressEnvelope'
+| 'PurolatorExpressEnvelope1030AM'
+| 'PurolatorExpressEnvelope9AM'
+| 'PurolatorExpressEnvelopeEvening'
+| 'PurolatorExpressEnvelopeInternational'
+| 'PurolatorExpressEnvelopeInternational1030AM'
+| 'PurolatorExpressEnvelopeInternational1200'
+| 'PurolatorExpressEnvelopeInternational9AM'
+| 'PurolatorExpressEnvelopeUS'
+| 'PurolatorExpressEnvelopeUS1030AM'
+| 'PurolatorExpressEnvelopeUS1200'
+| 'PurolatorExpressEnvelopeUS9AM'
+| 'PurolatorExpressEvening'
+| 'PurolatorExpressInternational'
+| 'PurolatorExpressInternational1030AM'
+| 'PurolatorExpressInternational1200'
+| 'PurolatorExpressInternational9AM'
+| 'PurolatorExpressPack'
+| 'PurolatorExpressPack1030AM'
+| 'PurolatorExpressPack9AM'
+| 'PurolatorExpressPackEvening'
+| 'PurolatorExpressPackInternational'
+| 'PurolatorExpressPackInternational1030AM'
+| 'PurolatorExpressPackInternational1200'
+| 'PurolatorExpressPackInternational9AM'
+| 'PurolatorExpressPackUS'
+| 'PurolatorExpressPackUS1030AM'
+| 'PurolatorExpressPackUS1200'
+| 'PurolatorExpressPackUS9AM'
+| 'PurolatorExpressUS'
+| 'PurolatorExpressUS1030AM'
+| 'PurolatorExpressUS1200'
+| 'PurolatorExpressUS9AM'
+| 'PurolatorGround'
+| 'PurolatorGround1030AM'
+| 'PurolatorGround9AM'
+| 'PurolatorGroundDistribution'
+| 'PurolatorGroundEvening'
+| 'PurolatorGroundRegional'
+| 'PurolatorGroundUS';
