+++ types/ServiceLevel/RRDonnelley.d.ts
@@ -1,0 +1,17 @@
+export declare type ServiceLevelRRDonnelley =
+| 'CourierServiceDDP'
+| 'CourierServiceDDU'
+| 'DomesticEconomyParcel'
+| 'DomesticParcelBPM'
+| 'DomesticPriorityParcel'
+| 'DomesticPriorityParcelBPM'
+| 'EMIService'
+| 'EconomyParcelService'
+| 'IPAService'
+| 'ISALService'
+| 'PMIService'
+| 'PriorityParcelDDP'
+| 'PriorityParcelDDU'
+| 'PriorityParcelDeliveryConfirmationDDP'
+| 'PriorityParcelDeliveryConfirmationDDU'
+| 'ePacketService';
