+++ types/ServiceLevel/RoyalMail.d.ts
@@ -1,0 +1,27 @@
+export declare type ServiceLevelRoyalMail =
+| 'InternationalSigned'
+| 'InternationalStandard'
+| 'InternationalTracked'
+| 'InternationalTrackedAndSigned'
+| '1stClass'
+| '1stClassSignedFor'
+| '2ndClass'
+| '2ndClassSignedFor'
+| 'RoyalMail24'
+| 'RoyalMail24SignedFor'
+| 'RoyalMail48'
+| 'RoyalMail48SignedFor'
+| 'SpecialDeliveryGuaranteed1pm'
+| 'SpecialDeliveryGuaranteed9am'
+| 'StandardLetter1stClass'
+| 'StandardLetter1stClassSignedFor'
+| 'StandardLetter2ndClass'
+| 'StandardLetter2ndClassSignedFor'
+| 'Tracked24'
+| 'Tracked24HighVolume'
+| 'Tracked24HighVolumeSignature'
+| 'Tracked24Signature'
+| 'Tracked48'
+| 'Tracked48HighVolume'
+| 'Tracked48HighVolumeSignature'
+| 'Tracked48Signature';
