+++ types/ServiceLevel/SpeeDee.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelSpeeDee = 'SpeeDeeDelivery';
