+++ types/ServiceLevel/SprintShip.d.ts
@@ -1,0 +1,4 @@
+export declare type ServiceLevelSprintShip =
+| 'RoutedDeliveries'
+| 'OnDemand'
+| 'WhiteGloveDeliveries';
