+++ types/ServiceLevel/StarTrack.d.ts
@@ -1,0 +1,4 @@
+export declare type ServiceLevelStarTrack =
+| 'StartrackExpress'
+| 'StartrackPremium'
+| 'StartrackFixedPricePremium';
