+++ types/ServiceLevel/TForce.d.ts
@@ -1,0 +1,5 @@
+export declare type ServiceLevelTForce =
+| 'SameDay'
+| 'SameDayWhiteGlove'
+| 'NextDay'
+| 'NextDayWhiteGlove';
