+++ types/ServiceLevel/UDS.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelUDS = 'DeliveryService';
