+++ types/ServiceLevel/UPS.d.ts
@@ -1,0 +1,13 @@
+export declare type ServiceLevelUPS =
+| 'Ground'
+| 'UPSStandard'
+| 'UPSSaver'
+| 'Express'
+| 'ExpressPlus'
+| 'Expedited'
+| 'NextDayAir'
+| 'NextDayAirSaver'
+| 'NextDayAirEarlyAM'
+| '2ndDayAir'
+| '2ndDayAirAM'
+| '3DaySelect';
