+++ types/ServiceLevel/UPSMailInnovations.d.ts
@@ -1,0 +1,6 @@
+export declare type ServiceLevelUPSMailInnovations =
+| 'First'
+| 'Priority'
+| 'ExpeditedMailInnovations'
+| 'PriorityMailInnovations'
+| 'EconomyMailInnovations';
