+++ types/ServiceLevel/USPS.d.ts
@@ -1,0 +1,11 @@
+export declare type ServiceLevelUSPS =
+| 'First'
+| 'Priority'
+| 'Express'
+| 'ParcelSelect'
+| 'LibraryMail'
+| 'MediaMail'
+| 'FirstClassMailInternational'
+| 'FirstClassPackageInternationalService'
+| 'PriorityMailInternational'
+| 'ExpressMailInternational';
