+++ types/ServiceLevel/Veho.d.ts
@@ -1,0 +1,1 @@
+export declare type ServiceLevelVeho = 'nextDay' | 'sameDay';
